Split platform out of core
