+++ simplecpreprocessor/core.py
@@ -1,52 +1,6 @@
-import platform
-from simplecpreprocessor import filesystem, tokens
+from simplecpreprocessor import filesystem, tokens, platform
+from simplecpreprocessor.exceptions import ParseError
 
-
-class ParseError(Exception):
-    pass
-
-
-def calculate_windows_constants(bitness=None):
-    if bitness is None:
-        bitness, _ = platform.architecture()
-    constants = {
-        "WIN32": "WIN32", "_WIN64": "_WIN64"}
-    if bitness == "64bit":
-        constants["WIN64"] = "WIN64"
-        constants["_WIN64"] = "_WIN64"
-    elif not bitness == "32bit":
-        raise Exception("Unsupported bitness %s" % bitness)
-    return constants
-
-
-def calculate_linux_constants(bitness=None):
-    if bitness is None:
-        bitness, _ = platform.architecture()
-    constants = {
-        "__linux__": "__linux__"
-    }
-    if bitness == "32bit":
-        constants["__i386__"] = "__i386__"
-    elif bitness == "64bit":
-        constants["__x86_64__"] = "__x86_64"
-    else:
-        raise Exception("Unsupported bitness %s" % bitness)
-    return constants
-
-
-def calculate_platform_constants():
-    system = platform.system()
-    if system == "Windows":
-        constants = calculate_windows_constants()
-    elif system == "Linux":
-        constants = calculate_linux_constants()
-    else:
-        raise ParseError("Unsupported platform %s" % platform)
-    constants["__SIZE_TYPE__"] = "size_t"
-    return constants
-
-
-PLATFORM_CONSTANTS = calculate_platform_constants()
 PRAGMA_ONCE = "pragma_once"
 IFDEF = "ifdef"
 IFNDEF = "ifndef"
@@ -57,7 +11,7 @@ class Preprocessor(object):
 
     def __init__(self, line_ending=tokens.DEFAULT_LINE_ENDING,
                  include_paths=(), header_handler=None,
-                 platform_constants=PLATFORM_CONSTANTS,
+                 platform_constants=platform.PLATFORM_CONSTANTS,
                  ignore_headers=()):
         self.ignore_headers = ignore_headers
         self.include_once = {}
@@ -258,7 +212,8 @@ class Preprocessor(object):
 
 
 def preprocess(f_object, line_ending="\n", include_paths=(),
-               header_handler=None, platform_constants=PLATFORM_CONSTANTS,
+               header_handler=None,
+               platform_constants=platform.PLATFORM_CONSTANTS,
                ignore_headers=()):
     r"""
     This preprocessor yields chunks of text that combined results in lines
